ignored atmel studio

+++ AAquad_firmware/AAquad_firmware/header_files/pass_to_pwm_chip.h
@@ -1,10 +1,12 @@
 #include <avr/io.h>
 
+#include <stdint.h>
+
 #ifndef PASS_TO_PWM_CHIP
 
 #define PASS_TO_PWM_CHIP
 
-int pass_to_pwm_chip(int* motors);
+int pass_to_pwm_chip(uint8_t* motors);
 
 
 #endif
